# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "D-Form API Documentation",
        "description": "Dokumentasi API lengkap untuk D-Form - Platform manajemen event dan recruitment DOSCOM",
        "contact": {
            "name": "DOSCOM - Dinus Open Source Community",
            "email": "doscom@dinus.ac.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Development Server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/participants": {
            "get": {
                "tags": [
                    "Participants"
                ],
                "summary": "Get all participants",
                "description": "Retrieve a list of all participants with optional event filter",
                "operationId": "b56cc0b1479de99c053898dea57d2a59",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "query",
                        "description": "Filter by event ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "is_presence",
                        "in": "query",
                        "description": "Filter by presence status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Participant"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Participants"
                ],
                "summary": "Register a new participant",
                "description": "Register a participant for an event",
                "operationId": "700a91d1e48b221e6d27eb86f50a4dc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id",
                                    "name",
                                    "school",
                                    "email"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "9d4e5f6a-7b8c-9d0e-1f2a-3b4c5d6e7f8a"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Smith"
                                    },
                                    "school": {
                                        "type": "string",
                                        "example": "Universitas Dian Nuswantoro"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Participant registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Participant registered successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Participant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/participants/{id}": {
            "get": {
                "tags": [
                    "Participants"
                ],
                "summary": "Get participant by ID",
                "description": "Retrieve a single participant by its ID",
                "operationId": "022e40e8979c7d7a7f156a8c80aeaa5f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Participant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Participant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Participants"
                ],
                "summary": "Update participant",
                "description": "Update participant information",
                "operationId": "c42492b70d7706860bae3c7f4966cfa4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Smith Updated"
                                    },
                                    "school": {
                                        "type": "string",
                                        "example": "Universitas Dian Nuswantoro"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane.updated@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Participant updated successfully"
                    },
                    "404": {
                        "description": "Participant not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Participants"
                ],
                "summary": "Delete participant",
                "description": "Delete a participant registration",
                "operationId": "2a2003f3c1c8a6b6164344f3bd5085ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participant deleted successfully"
                    },
                    "404": {
                        "description": "Participant not found"
                    }
                }
            }
        },
        "/api/participants/{id}/presence": {
            "patch": {
                "tags": [
                    "Participants"
                ],
                "summary": "Mark participant presence",
                "description": "Mark a participant as present for an event",
                "operationId": "0f530b00825b1e97749148e0f27bed1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Participant ID (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Presence marked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Presence marked successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Participant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Participant not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Participant": {
                "title": "Participant",
                "description": "Participant model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9d4e5f6a-7b8c-9d0e-1f2a-3b4c5d6e7f8a"
                    },
                    "event_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9d4e5f6a-7b8c-9d0e-1f2a-3b4c5d6e7f8a"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jane Smith"
                    },
                    "school": {
                        "type": "string",
                        "example": "Universitas Dian Nuswantoro"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "jane@example.com"
                    },
                    "is_presence": {
                        "type": "boolean",
                        "example": false
                    },
                    "presence_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-08T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-08T12:00:00.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}